requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
schedule>=1.2.0
colorama>=0.4.6
redis>=5.0.0
web3>=6.0.0
eth-account>=0.9.0
mnemonic>=0.20

